requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pymongo>=4.11.0
voyageai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
boto3>=1.35.0
python-multipart>=0.0.9
